# controllers/order_controller.py
from flask import Blueprint, request, jsonify, Response, g
from service import order_service
from dto.order_dto import OrderResponseDTO, OrderPageDTO, CreateOrderRequestDTO
from decorator.tokenDecorator import token_required
from decorator.logRequestDecorator import log
from datetime import datetime
from helpers.ApiExceptions import APIException
import json
import requests
import boto3
import logging


logger = logging.getLogger(__name__)

order_bp = Blueprint('order_bp', __name__)

lambda_client = boto3.client('lambda', region_name = 'us-east-1')

@order_bp.route('/orders/stats/lambda', methods=['GET'])
@token_required(roles=['artist', 'internal_service'])
@log('../logs/ficherosalida.log')
def get_statistics_orders():
    """
    Invoca Lambda a través de API Gateway (funciona en cuentas educativas)
    """
    try:
        days = request.args.get('days', 7, type=int)
        username = request.user_claims.get('username')
        
        api_gateway_url = "https://jf9o4ntpik.execute-api.us-east-1.amazonaws.com/default/OrderStatisticsFunction"
        
        # Llamar a API Gateway
        response = requests.get(api_gateway_url, params={
            'days': days,
            'username': username
        })
        
        if response.status_code == 200:
            return jsonify(response.json()), 200
        else:
            return jsonify({
                'success': False,
                'error': 'Error en Lambda via API Gateway',
                'status_code': response.status_code
            }), 500
            
    except Exception as e:
        logger.error(f"Error invocando Lambda via API: {str(e)}")
        return jsonify({
            'success': False,
            'error': 'Servicio de estadísticas temporalmente no disponible',
            'fallback_mode': 'active'
        }), 503

@order_bp.route('/orders/<string:orderId>', methods=['GET'])
@token_required(roles=['artist', 'internal_service'])  #-> Validate user token
@log('../logs/ficherosalida.log')
def proccess_orders_by_id(orderId: str):
    try:
        # Llamo al servicio para buscar compras que coincidan con orderId
        order = order_service.OrderService.find_order(orderId)
        if order is None:

            raise APIException(
                message=f"La orden {orderId} no existe",
                status=404,
                error="Not Found"
            )
        
        order_dict = order_service.OrderService.order_to_dict(order)
        # Validate the entry Order with Pydantic. Converting SQLAlchemy obj -> Pydantic DTO
        orderDTO = OrderResponseDTO.model_validate(order_dict)

        return Response(
            orderDTO.model_dump_json(),
            status=200,
            mimetype='application/json'
        )

    except ValueError as e:
        logger.warning(f"Error de validación: {str(e)}")
        return jsonify({'message': str(e)}), 400
    
@order_bp.route('/orders/<string:orderId>', methods=['PATCH'])
@token_required(roles=['artist', 'internal_service']) #-> Validate user token
@log('../logs/ficherosalida.log')
def update_order_by_id(orderId: str):
    try:
        data = request.get_json() # Obtenemos los datos del cuerpo de la consulta
                
        if not data:

            raise APIException(
                message='Datos de actualización requeridos',
                status=400,
                error='Bad Request'
            )
        
        is_updatable = order_service.OrderService.is_order_updatable(orderId)

        if not is_updatable:

            raise APIException(
                message=f"La orden {orderId} no exite o no se puede actualizar (ya fue enviada, cancelada o entregada)",
                status=400,
                error='Bad Request'
            )
        
        order_update = order_service.OrderService.update_order(orderId, data)

        if not order_update:

            raise APIException(
                message=f"La compra {orderId} no existe",
                status=404,
                error='Not Found'
            )

        order_updated_dict = order_service.OrderService.order_to_dict(order_update)
        orderDTO_updated = OrderResponseDTO.model_validate(order_updated_dict)

        return Response(
            response = orderDTO_updated.model_dump_json(),
            status = 200,
            mimetype = 'application/json'
        )   
        
    except ValueError as e:
        logger.warning(f"Error de validación: {str(e)}")
        return jsonify({'error': str(e)}), 400   
    
@order_bp.route('/orders/<string:orderId>', methods=['DELETE'])
@token_required(roles=['artist', 'internal_service']) #-> Validate user token
@log('../logs/ficherosalida.log')   
def delete_order_by_id(orderId: str):
    try: 
        deleted = order_service.OrderService.delete(orderId)

        if not deleted:

            raise APIException(
                message=f"La compra {orderId} no existe",
                status=404,
                error='Not Found'
            )

        return Response(
            response="Compra eliminada correctamente",
            status=204,
            mimetype='application/json'
        )
        
    except ValueError as e:
        logger.warning(f"Error de validación: {str(e)}")
        return jsonify({'error': str(e)}), 400   


@order_bp.route('/orders', methods=['GET'])
@token_required(roles=['artist', 'internal_service'])    #-> Validate user token
@log('../logs/ficherosalida.log')
def procesar_compras():
    filters = {}
    orders_tuple = None
    try:
        if request.method == 'GET': # Save the query params of the HTTP request in a dictionary
            
            filters['seller'] = request.args.get('seller')
            filters['status'] = request.args.get('status')
            filters['dateFrom'] = request.args.get('dateFrom')
            filters['dateTo'] = request.args.get('dateTo')
            filters['page'] = request.args.get('page', 0, type=int)
            filters['size'] = request.args.get('size', 20, type=int)

            # Paginating validation
            if filters['page'] < 0:
                
                raise APIException(
                    message='La pagina debe de ser mayor o igual a 0',
                    status=400,
                    error='Bad Request'
                )
            
            if filters['size'] < 20 or filters['size'] > 100:
                
                raise APIException(
                    message='El tamaño de la página debe de ser mayor o igual a 20',
                    status=400,
                    error='Bad Request'
                )

            # CASE: No filter applied
            if filters is None: 
                orders_tuple = order_service.OrderService.list_orders(filters['size'], filters['page'])
            else:
                orders_tuple = order_service.OrderService.list_orders_with_filters(filters)    

            # Access to the Tuple elements of orders
            order_list = orders_tuple[0]
            total_elements = orders_tuple[1]
            total_pages = orders_tuple[2]

            # Convert orders into Page DTO
            pageDTO = OrderPageDTO(
                content = [OrderResponseDTO.from_orm(order) for order in order_list],
                totalElements = total_elements,
                totalPages = total_pages,
                page = filters['page'],
                size = filters['size']
            )    

            return Response(
                pageDTO.model_dump_json(),
                status=200,
                mimetype='application/json'
            )
            
    except ValueError as e:
        return jsonify({
            'message': str(e),
            'status': 400,
            'error': 'Bad Request'
        }), 400

@order_bp.route('/orders', methods=['POST'])
@token_required(roles=['artist', 'internal_service']) #-> Validate user token
@log('../logs/ficherosalida.log')
def create_order():
    try:
        # Obtain data from the request
        data = request.get_json()

        if not data:

            raise APIException(
                message='Datos en json no requeridos',
                status=400,
                error='Bad Request'
            )
        
        # Validate data request with Pylantic
        data_validated = CreateOrderRequestDTO.model_validate(data)

        # Get user from JWT
        current_user = request.user_claims.get('username')

        order_created = order_service.OrderService.save(
            data_validated, 
            current_user
        )
        
        # Converto to DTO the response
        #order_created_dto = OrderResponseDTO.from_orm(order_created)
        response = {
            "purchaseId": order_created.public_id,
            "amount": order_created.total
        }

        return Response(
            #order_created_dto.model_dump_json(),
            response=json.dumps(response),
            status=201,
            mimetype='application/json'
        )
        
    except order_service.ProductNotFoundException as e:
        return jsonify({
            'error': 'Producto no encontrado',
            'details': str(e)
        }), 404
        
    except ValueError as e:
        return jsonify({
            'error': 'Error de validación',
            'details': str(e)
        }), 400
        
    
@order_bp.route('/orders/<string:orderId>/confirm', methods=['POST'])
@token_required(roles=['artist', 'internal_service']) # -> Validate user token
@log('../logs/ficherosalida.log')
def confirm_order(orderId: str):
    """
    Confirma una compra: verifica stock → comprueba pago → actualiza stock
    """
    try:
        # 1. Verificar que la orden existe
        order = order_service.OrderService.find_order(orderId)

        if not order:

            raise APIException(
                message=f'La orden {orderId} no existe',
                status=404,
                error='Not Found'
            )
            
        
        # 2. Validar que la orden esté en estado procesable
        if not order_service.OrderService.is_order_confirm(orderId):

            raise APIException(
                message=f'La orden {orderId} no puede ser confirmada',
                status=400,
                error='Bad Request'
            )
        
        # 3. Verificar disponibilidad de stock ANTES del pago
        stock_availability = order_service.OrderService.check_stock_availability(orderId)

        if not stock_availability['all_available']:

            raise APIException(
                message=f"Stock insuficiente para algunos productos",
                status=409,
                error='Conflict'
            )
                
        # 4. Actualizar stock de los productos de la compra de forma lógica
        stock_update_result = order_service.OrderService.update_product_stock(orderId, False) 
        
        # 5. Obtener datos de pago del request
        data = request.get_json()

        if not data:

            raise APIException(
                message='Datos de pago requeridos',
                status=400,
                error='Bad Request'
            )
        
        # 6. Procesar pago
        payment_result = order_service.OrderService.process_order_payment(orderId, data)

        # 7. Obtenemos la información del pedido actualizada
        order_info = payment_result['transaction_data']
        
        if not order_info['success']:

            raise APIException(
                message=f'Error en el procesamiento del pago',
                status=402,
                error='Payment Error'
            )
        

        # 8. Comprobar si se ha realizado el pago del pedido para actualizar el stock
        if not order_service.OrderService.is_order_paid(orderId):

            order_service.OrderService.update_product_stock(orderId, True) # Reestablecemos el stock

            raise APIException(
                message='No se ha realizaod pago y no se puede actualizar el stock',
                status=409,
                error='Conflict'
            )

        
        
        """if not stock_update_result['success']:
            # Revertir el pago si la actualización de stock falla
            order_service.OrderService.reverse_payment(orderId)
            return jsonify({
                'error': 'Error al actualizar el stock',
                'details': stock_update_result.get('details', 'Error desconocido')
            }), 500"""
        
        # 7. Comprobar que se ha actualizado correctamente el stock para terminar la confirmación
        """if stock_update_result.get('success') is True:
            order_dict = order_service.OrderService.order_to_dict(order)
            orderDTO = OrderResponseDTO.model_validate(order_dict)
            
            return Response(
                orderDTO.model_dump_json(),
                status=200,
                mimetype='application/json'
            )
        else: # Error return

            return Response(
                response = stock_update_result.get('message'),
                status = 409
            )"""
        
    except order_service.OrderNotFoundException as e:
        return jsonify({'error': str(e)}), 404
    except order_service.PaymentProcessingException as e:
        return jsonify({'error': 'Error en el procesamiento del pago', 'details': str(e)}), 502