# swagger/swagger.yaml
openapi: "3.0.1"
info:
  title: Servicio de Compras
  description: |-
    API de Undersounds para la gestión integral de compras, permitiendo la 
    administración de pedidos, control de inventario y consulta del catálogo de productos.
  license:
    name: Apache License 2.0
    url: https://www.apache.org/licenses/LICENSE-2.0.html
  version: 1.0.0
servers:
  - url: /api/v1/ #https://undersounds.com/api/v1 --> Produccion
    description: Servidor local de desarrollo
  - url: https://undersounds.com/api
    description: Servidor de produccion
...
paths:
  /orders:
    get:
      tags:
        - order
      summary: Listar compras con filtros
      operationId: "controllers.order_controller.procesar_compras"
      #operationId: "listOrders"
      description: Obtiene una lista de compras filtradas por cualquier atributo.
                  | Si no se especifican filtros, se listan todas las compras.
                  | Requiere JWT.
      security:
        - bearerAuth: [ ]
      parameters:
        - in: query
          name: seller
          schema:
            type: string
          description: Filtra por nombre de usuario del vendedor. Si no se proporciona, se devuelven las compras de todos los vendedores.
          example: "coldplay_official"
        - in: query
          name: status
          schema:
            $ref: "#/components/schemas/OrderStatus"
          description: Filtra por estado de la compra. Si no se proporciona, se devuelven las compras sin tener en cuenta el estado.
          example: "PAID"
        - in: query
          name: dateFrom
          schema:
            type: string
            format: date
          description: Fecha mínima de la compra. Si no se proporciona, se devuelven las compras desde la primera.
          example: "2025-10-01"
        - in: query
          name: dateTo
          schema:
            type: string
            format: date
          description: Fecha máxima de la compra. Si no se proporciona, se devuelven las compras hasta la última.
          example: "2025-10-17"
        - in: query
          name: page
          schema:
            type: integer
            default: 0
          description: Número de página.
          example: 0
        - in: query
          name: size
          schema:
            type: integer
            default: 20
          description: Tamaño de página.
          example: 20
      responses:
        '200':
          description: Lista paginada de compras.
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/OrdersPage'
        '400':
          description: Parámetros de paginacion inválidos
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
              examples:
                invalidPage:
                  summary: Página inválida
                  value:
                    code: 400
                    message: "La pagina debe de ser mayor o igual a 0"
                invalidSize:
                  summary: Tamaño inválido
                  value:
                    code: 400
                    message: "El tamaño de la página debe de ser mayor o igual a 20"     
        '401':
          $ref: '#/components/responses/UnauthorizedError'
        '500':
          description: Error interno del Servidor
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'                       

    post:
      tags:
        - order
      summary: Crear una nueva compra
      operationId: "controllers.order_controller.create_order"
      #operationId: "createOrder"
      description: |
        Permite crear una compra con los productos indicados.
        La compra se crea con el estado PENDING.
        Requiere JWT.
      security:
        - bearerAuth: [ ]
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/CreateOrderRequest'
            example:
              singleItem:
                summary: Compra con un producto
                value:
                  items:
                    - productId: "550e8400-e29b-41d4-a716-446655440000"  
                      quantity: 2
              multipleItems:
                summary: Compra de varios productos
                value:
                  items:
                    - productId: "550e8400-e29b-41d4-a716-446655440000"
                      quantity: 2
                    - productId:   "660e8400-e29b-41d4-a716-446655440001"
                      quantity: 3     

      responses:
        '201':
          description: Compra creada correctamente.
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/OrderResponse'
        '400':
          description: Error en el body de la petición.
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
              examples:
                errorQuantity:
                  summary: Cantidad incorrecta.
                  value:
                    code: 400
                    message: "Solo existen 2 unidades para el artículo 550e8400-e29b-41d4-a716-446655440000 y se han especificado 3."
        '401':
          $ref: '#/components/responses/UnauthorizedError'
        '404':
          description: Algunos de los items no se encontraron.
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
              examples:
                notFound:
                  summary: Contenido no encontrado
                  value:
                    code: 404
                    message: "El artículo 550e8400-e29b-41d4-a716-446655440000 no existe"
        '500':
          description: Error interno del Servicio
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'            
  /orders/{orderId}:
    get:
      tags:
        - order
      summary: Obtener una compra
      operationId: "controllers.order_controller.proccess_orders_by_id"
      #operationId: "getOrdersById"
      description: Recupera los datos de una compra por su ID público. Requiere JWT.
      security:
        - bearerAuth: [ ]
      parameters:
        - in: path
          name: orderId
          required: true
          description: ID público de la compra.
          schema:
            type: string
            format: uuid
            example: "550e8400-e29b-41d4-a716-446655440000"
      responses:
        '200':
          description: Datos de la compra.
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/OrderResponse'
        '401':
          $ref: '#/components/responses/UnauthorizedError'
        '404':
          $ref: '#/components/responses/OrderNotFoundError'
        '500':
          description: Error interno del servidor.
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'  

    patch:
      tags:
        - order
      summary: Actualizar algún atributo de una compra
      operationId: "controllers.order_controller.update_order_by_id"
      #operationId: "updateOrdersById"
      description: |
        Actualiza el estado de una orden. Transiciones permitidas:
          - PENDING → PAID, CANCELLED
          - PAID → SHIPPED, CANCELLED (requiere reembolso)
          - SHIPPED → No se puede modificar
          - CANCELLED → Estado final.
        Requiere JWT o rol administrador.
      security:
        - bearerAuth: [ ]
        - adminAuth: [ ]
      parameters:
        - in: path
          name: orderId
          required: true
          description: ID público de la compra.
          schema:
            type: string
            format: uuid
            example: "550e8400-e29b-41d4-a716-446655440000"
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              properties:
                status:
                  $ref: "#/components/schemas/OrderStatus"
                  description: Nuevo estado de la compra.
      responses:
        '200':
          description: Compra actualizada correctamente.
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/OrderResponse'
        '401':
          $ref: '#/components/responses/UnauthorizedError'
        '404':
          $ref: '#/components/responses/OrderNotFoundError'
        '500':
          description: Error interno del servidor.
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'  

    delete:
      tags:
        - order
      summary: Eliminar una compra
      operationId: "controllers.order_controller.delete_order_by_id"
      #operationId: "deleteOrdersById"
      description: Si su estado aún está PENDIENTE el usuario puede cancelarla, de lo contrario solo administradores pueden eliminarla.
      security:
        - bearerAuth: [ ]
        - adminAuth: [ ]
      parameters:
        - in: path
          name: orderId
          description: Identificador público de la compra a eliminar/cancelar.
          required: true
          schema:
            type: string
            format: uuid
            example: "550e8400-e29b-41d4-a716-446655440000"
      responses:
        '204':
          description: Compra eliminada correctamente.
          content:
            application/json:
              schema:
                type: object
                properties:
                  message:
                    type: string
                    example: "Compra eliminada correctamente"
        '401':
          $ref: '#/components/responses/UnauthorizedError'
        '404':
          $ref: '#/components/responses/OrderNotFoundError'
        '500':
          description: Error interno del servidor.
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'  
  /orders/{orderId}/confirm:
    post:
      tags:
        - order
      summary: Confirmar y procesar una compra
      operationId: "controllers.order_controller.confirm_order"
      description: |
        Confirma una compra realizando las siguientes operaciones:
        1. Verifica que la orden exista y esté en estado PENDING
        2. Verifica disponibilidad de stock para todos los productos
        3. Procesa el pago a través del servicio de pagos
        4. Actualiza el stock de los productos (con rollback si falla)
        5. Marca la orden como PAID
        
        Requiere JWT.
      security:
        - bearerAuth: [ ]
      parameters:
        - in: path
          name: orderId
          required: true
          description: ID público de la compra a confirmar.
          schema:
            type: string
            format: uuid
            example: "550e8400-e29b-41d4-a716-446655440000"
      requestBody:
        required: true
        description: Datos necesarios para procesar el pago
        content:
          application/json:
            schema:
              type: object
              required:
                - payment_method
              properties:
                payment_method:
                  type: string
                  enum:
                    - PLATFORM_BALANCE
                    - CREDIT_CARD
                    - PAYPAL
                  description: Método de pago a utilizar
                  example: "PLATFORM_BALANCE"
            examples:
              platformBalance:
                summary: Pago con saldo de la plataforma
                value:
                  payment_method: "PLATFORM_BALANCE"
              creditCard:
                summary: Pago con tarjeta de crédito
                value:
                  payment_method: "CREDIT_CARD"
      responses:
        '200':
          description: Compra confirmada y procesada exitosamente.
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/OrderResponse'
        '400':
          description: La orden no puede ser confirmada o faltan datos de pago.
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
              examples:
                notConfirmable:
                  summary: Orden no confirmable
                  value:
                    error: "La orden 550e8400-e29b-41d4-a716-446655440000 no puede ser confirmada"
                missingPaymentData:
                  summary: Datos de pago faltantes
                  value:
                    error: "Datos de pago requeridos"
        '401':
          $ref: '#/components/responses/UnauthorizedError'
        '402':
          description: Error en el procesamiento del pago.
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
              examples:
                paymentFailed:
                  summary: Pago fallido
                  value:
                    error: "Error en el procesamiento del pago"
                    details: "Saldo insuficiente"
        '404':
          $ref: '#/components/responses/OrderNotFoundError'
        '409':
          description: Stock insuficiente para algunos productos.
          content:
            application/json:
              schema:
                type: object
                properties:
                  error:
                    type: string
                    example: "Stock insuficiente para algunos productos"
                  details:
                    type: array
                    items:
                      type: object
                      properties:
                        product_id:
                          type: string
                        available:
                          type: boolean
                        current_stock:
                          type: integer
                        required:
                          type: integer
              examples:
                insufficientStock:
                  summary: Stock insuficiente
                  value:
                    error: "Stock insuficiente para algunos productos"
                    details:
                      - product_id: "550e8400-e29b-41d4-a716-446655440000"
                        available: false
                        current_stock: 1
                        required: 3
        '500':
          description: Error interno del servidor.
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
        '502':
          description: Error comunicándose con servicios externos (pagos, inventario).
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
              examples:
                paymentServiceDown:
                  summary: Servicio de pagos no disponible
                  value:
                    error: "Error en el procesamiento del pago"
                    details: "Servicio de pagos no disponible"     
  /orders/stats/lambda:
    get:
      tags:
        - order
      summary: Obtener estadísticas de microservicio a través de AWS
      operationId: controllers.order_controller.get_statistics_orders
      description: |
        Obtiene las estadísticas generales del microservicio:  
        1. Realiza una llamada REST al desencadenador API Gateway  
        2. El API Gateway ejecuta una función Lambda AWS  
        3. Retorno de datos obtenidos por la función Lambda AWS

        Requiere JWT.
      security:
        - bearerAuth: []
      parameters:
        - in: query
          name: days
          required: true
          description: Instante de día en el que se quieren obtener las estadísticas
          schema:
            type: integer
            format: int64
            example: 5
      responses:
        '200':
          description: Estadísticas recuperadas desde AWS correctamente
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/LambdaResponse'
        '401':
          $ref: '#/components/responses/UnauthorizedError'
        '400':
          description: Error en los parámetros de la request
          content:
            application/json:
              schema:
                type: object
                properties:
                  code:
                    type: integer
                    example: 400
                  message:
                    type: string
                    example: "Error al enviar parámetros"
        '500':
          description: Error interno del servidor
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
components:
  securitySchemes:
    bearerAuth:
      type: http
      scheme: bearer
      bearerFormat: JWT
      description: Token JWT obtenido tras autenticación exitosa

    adminAuth:
      type: http
      scheme: bearer
      bearerFormat: JWT  
      description: Token JWT con rol de administrador
  schemas:
...
    OrderStatus:
      type: string
      description: |
        Estados de la orden y transiciones permitidas:
        - PENDING → PAID, CANCELLED
        - PAID → SHIPPED, CANCELLED (con reembolso)
        - SHIPPED → No se permiten más cambios
        - CANCELLED → Estado final
      enum:
        - PENDING
        - PAID
        - SHIPPED
        - CANCELLED
      example: PAID
...
    OrderResponse:
      type: object
      description: Respuesta ofrecida al cliente con información de una sola compra.
      required:
        - publicId
        - madeBy
        - items
        - createdAt
        - status
        - total
      properties:
        publicId:
          type: string
          format: uuid
          description: Identificador público único de la compra.
          example: "550e8400-e29b-41d4-a716-446655440000"
        madeBy:
          $ref: '#/components/schemas/Buyer'
          description: Usuario que ha realizado la compra.
        items:
          type: array
          items:
            $ref: '#/components/schemas/OrderItem'
          description: Productos que se han comprado.
        createdAt:
          type: string
          format: date-time
          description: Fecha en la que se realizó la compra.
          example: "2025-10-17T14:30:00Z"
        status: 
          $ref: '#/components/schemas/OrderStatus'
          description: Estado de la compra.
        total:
          type: number
          format: float
          description: Coste total en euros.
          example: 19.99
    CreateOrderRequest:
      type: object
      description: Petición para crear una compra.
      required:
        - items
      properties:
        items:
          type: array
          description: Productos a comprar con sus cantidades.
          minItems: 1
          maxItems: 50
          items:
            type: object
            required:
              - productId
              - quantity
            properties:
              productId:
                type: string
                format: uuid
                description: Identificador público del producto.
                example: "550e8400-e29b-41d4-a716-446655440000"
              quantity:
                type: integer
                minimum: 1
                maximum: 100
                description: Cantidad de unidades de este producto.
                example: 2
    OrdersPage:
      type: object
      description: Página ofrecida al cliente con información de las compras.
      properties:
        content:
          type: array
          items:
            $ref: "#/components/schemas/OrderResponse"
          description: Compras filtradas.
        page:
          type: integer
          default: 0
          minimum: 0
          example: 0
          description: Número de página que se quiere consultar.
        size:
          type: integer
          default: 20
          maximum: 100
          example: 20
          description: Tamaño de la página (máximo 100 registros).
        totalElements:
          type: integer
          minimum: 0
          example: 125
          description: Número de compras obtenidas.
        totalPages:
          type: integer
          minimum: 0
          example: 7
          description: Número de páginas totales del tamaño indicado.
    ErrorResponse:
      type: object
      description: Objeto con información del error en la petición.
      required:
        - error
      properties:
        error:
          type: string
          description: Mensaje descriptivo del error.
          example: "El campo 'items' es obligatorio."
        details:
          oneOf:
            - type: string
            - type: array
              items:
                type: string
            - type: object
          description: Información adicional sobre el error (puede ser string, array u objeto).
          example: "La cantidad debe ser mayor a 0"
  responses:
    UnauthorizedError:
      description: Usuario no autenticado.
      content:
        application/json:
          schema:
            $ref: '#/components/schemas/ErrorResponse'
          examples:
            unauthorized:
              summary: JWT inválido o ausente
              value:
                code: 401
                message: "Token requerido para realizar la operacion"
    OrderNotFoundError:
      description: Compra no encontrada.
      content:
        application/json:
          schema:
            $ref: '#/components/schemas/ErrorResponse'
          examples:
            notFound:
              summary: Compra no encontrado
              value:
                code: 404
                message: "La compra 550e8400-e29b-41d4-a716-446655440000 no existe"
security:
  - bearerAuth: []          